Worked on StudentSignOut and created a separate sign-in page for administrators and students.
Need to add:
 - Proper save function for checkboxes in tableview
 - Notification icon/view
 - Export/Import .CSV
