+++ src/main/java/Controllers/TransactionController.java
@@ -25,7 +25,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class TransactionSceneController implements Initializable {
+public class TransactionController implements Initializable {
 
 
     /*--------------------------------------------------------------------------------*/
