--- src/main/resources/Config-Files/config.properties
@@ -1,4 +1,0 @@
-#Sun Jan 14 20:57:19 EST 2024
-allBox=true
-lowQuantityBox=true
-lateSignOutBox=true
