+++ src/main/resources/Config-Files/config.settings
@@ -1,0 +1,4 @@
+#Mon Jan 15 18:31:48 EST 2024
+allBox=false
+lowQuantityBox=true
+lateSignOutBox=false
