+++ src/main/resources/com/example/orderly_inventory_management/SettingScene.fxml
@@ -12,9 +12,10 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.shape.Line?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="parent" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="768.0" prefWidth="1366.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.SettingSceneController">
+<AnchorPane fx:id="parent" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="768.0" prefWidth="1366.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.SettingController">
    <children>
       <BorderPane style="-fx-background-color: #101010;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
          <left>
@@ -37,14 +38,6 @@
                      </cursor>
                   </Button>
                   <Button fx:id="dashboardButton" alignment="BASELINE_LEFT" layoutX="15.0" layoutY="133.0" mnemonicParsing="false" onAction="#switchToDashboard" prefHeight="50.0" stylesheets="@../../../Stylesheets/Global.css" text="Dashboard" textFill="#6c757d" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="60.0">
-                     <font>
-                        <Font name="System Bold" size="18.0" />
-                     </font>
-                     <cursor>
-                        <Cursor fx:constant="HAND" />
-                     </cursor>
-                  </Button>
-                  <Button fx:id="notiButton" alignment="BASELINE_LEFT" layoutX="15.0" layoutY="283.0" mnemonicParsing="false" prefHeight="50.0" stylesheets="@../../../Stylesheets/Global.css" text="Notifications" textFill="#6c757d" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="210.0">
                      <font>
                         <Font name="System Bold" size="18.0" />
                      </font>
@@ -91,6 +84,9 @@
                      <image>
                         <Image url="@../../../Assets/AppIcon.png" />
                      </image>
+                     <effect>
+                        <ColorAdjust brightness="1.0" />
+                     </effect>
                   </ImageView>
                   <Label alignment="BASELINE_LEFT" layoutX="703.0" layoutY="7.0" text="ORDERLY" textFill="WHITE" AnchorPane.leftAnchor="40.0">
                      <font>
@@ -131,7 +127,6 @@
                         <Font name="System Bold" size="18.0" />
                      </font>
                   </Button>
-                  <ComboBox layoutX="63.0" layoutY="55.0" prefWidth="150.0" promptText="Default Dark" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.leftAnchor="30.0" AnchorPane.topAnchor="65.0">
                   <ComboBox layoutX="63.0" layoutY="55.0" prefWidth="150.0" promptText="Default Dark" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="60.0">
                      <effect>
                         <ColorAdjust brightness="1.0" />
@@ -149,13 +144,9 @@
                   </Label>
                   <Line endX="50.0" endY="-92.0" layoutX="499.0" layoutY="92.0" startX="50.0" startY="595.5" stroke="#101010" AnchorPane.bottomAnchor="10.0" AnchorPane.topAnchor="10.0" />
                   <Label layoutX="15.0" layoutY="425.0" text="Accounts" textFill="WHITE" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="425.0">
-                  <Label layoutX="15.0" layoutY="246.0" text="Export .CSV Directory:" textFill="WHITE" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="245.0" />
-                  <TextField layoutX="196.0" layoutY="241.0" promptText="C://User/Downloads" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.leftAnchor="135.0" AnchorPane.topAnchor="241.0" />
-                  <Label layoutX="29.0" layoutY="294.0" text="Miscellaneous" textFill="WHITE" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="300.0">
                      <font>
                         <Font name="System Bold" size="24.0" />
-                     </font>
-                  </Label>
+                     </font></Label>
                   <Label layoutX="33.0" layoutY="254.0" text="All reminders" textFill="WHITE" AnchorPane.leftAnchor="30.0" AnchorPane.topAnchor="250.0">
                      <font>
                         <Font size="15.0" />
@@ -179,7 +170,7 @@
                      </font>
                   </Label>
                   <Button layoutX="465.0" layoutY="473.0" mnemonicParsing="false" stylesheets="@../../../Stylesheets/Global.css" text="Change" AnchorPane.rightAnchor="650.0" AnchorPane.topAnchor="475.0" />
-                  <Button layoutX="449.0" layoutY="513.0" mnemonicParsing="false" stylesheets="@../../../Stylesheets/Global.css" text="Change" />
+                  <Button layoutX="449.0" layoutY="513.0" mnemonicParsing="false" stylesheets="@../../../Stylesheets/Global.css" text="Change" AnchorPane.rightAnchor="650.0" AnchorPane.topAnchor="515.0" />
                   <Line endX="433.0" layoutX="115.0" layoutY="417.0" startX="-100.0" stroke="#101010" />
                   <CheckBox fx:id="allBox" layoutX="490.0" layoutY="252.0" mnemonicParsing="false" stylesheets="@../../../Stylesheets/Global.css" />
                   <CheckBox fx:id="lateSignOutBox" layoutX="490.0" layoutY="212.0" mnemonicParsing="false" stylesheets="@../../../Stylesheets/Global.css" />
