+++ src/main/resources/com/example/orderly_inventory_management/StudentLoginScene.fxml
@@ -2,23 +2,23 @@
 
 <?import javafx.scene.Cursor?>
 <?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.effect.ColorAdjust?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="root" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="520.0" style="-fx-background-color: #FAF9F6;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.LoginSceneController">
+<AnchorPane fx:id="root" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="520.0" style="-fx-background-color: #fdfffc;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.StudentLoginController">
    <children>
-      <Label alignment="CENTER" contentDisplay="CENTER" layoutX="11.0" layoutY="99.0" text="ORDERLY" textAlignment="CENTER" textFill="#50576d" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="105.0">
+      <Label alignment="CENTER" contentDisplay="CENTER" layoutX="11.0" layoutY="99.0" text="ORDERLY" textAlignment="CENTER" textFill="#3a405a" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="105.0">
          <font>
             <Font name="Arial Bold" size="18.0" />
          </font>
       </Label>
-      <TextField fx:id="usernameField" layoutX="36.0" layoutY="173.0" prefHeight="30.0" prefWidth="240.0" promptText="Username" style="-fx-background-color: transparent; -fx-border-color: #50576D; -fx-border-width: 0px 0px 1.5px 0px;" AnchorPane.leftAnchor="135.0" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="160.0">
+      <TextField fx:id="usernameField" layoutX="36.0" layoutY="173.0" prefHeight="30.0" prefWidth="240.0" promptText="Username" style="-fx-background-color: transparent; -fx-border-color: #3a405a; -fx-border-width: 0px 0px 1.5px 0px;" AnchorPane.leftAnchor="135.0" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="160.0">
          <font>
             <Font name="Microsoft Sans Serif" size="12.0" />
          </font>
@@ -26,7 +26,7 @@
             <Cursor fx:constant="HAND" />
          </cursor>
       </TextField>
-      <Button fx:id="loginButton" layoutX="148.0" layoutY="287.0" maxHeight="-Infinity" minHeight="-Infinity" mnemonicParsing="false" onAction="#login" prefHeight="30.0" prefWidth="123.0" stylesheets="@../../../Stylesheets/Login.css" text="Sign In" textFill="WHITE" AnchorPane.leftAnchor="135.0" AnchorPane.topAnchor="265.0">
+      <Button fx:id="loginButton" layoutX="148.0" layoutY="287.0" maxHeight="-Infinity" minHeight="-Infinity" mnemonicParsing="false" onAction="#login" prefHeight="30.0" prefWidth="120.0" stylesheets="@../../../Stylesheets/Login.css" text="Sign In" textFill="WHITE" AnchorPane.leftAnchor="135.0" AnchorPane.topAnchor="255.0">
          <font>
             <Font name="System Bold" size="14.0" />
          </font>
@@ -34,7 +34,7 @@
             <Cursor fx:constant="HAND" />
          </cursor>
       </Button>
-      <Button fx:id="quitButton" layoutX="267.0" layoutY="287.0" maxHeight="-Infinity" minHeight="-Infinity" mnemonicParsing="false" onAction="#quitOnAction" prefHeight="30.0" prefWidth="123.0" stylesheets="@../../../Stylesheets/Login.css" text="Quit" textFill="WHITE" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="265.0">
+      <Button fx:id="quitButton" layoutX="267.0" layoutY="287.0" maxHeight="-Infinity" minHeight="-Infinity" mnemonicParsing="false" onAction="#quitOnAction" prefHeight="30.0" prefWidth="120.0" stylesheets="@../../../Stylesheets/Login.css" text="Quit" textFill="WHITE" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="255.0">
          <font>
             <Font name="System Bold" size="14.0" />
          </font>
@@ -42,7 +42,7 @@
             <Cursor fx:constant="HAND" />
          </cursor>
       </Button>
-      <PasswordField fx:id="passwordField" layoutX="36.0" layoutY="213.0" prefHeight="30.0" prefWidth="240.0" promptText="Password" style="-fx-background-color: transparent; -fx-border-color: #50576D; -fx-border-width: 0px0px1.5px0px;" AnchorPane.leftAnchor="135.0" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="200.0">
+      <PasswordField fx:id="passwordField" layoutX="36.0" layoutY="213.0" prefHeight="30.0" prefWidth="240.0" promptText="Password" style="-fx-background-color: transparent; -fx-border-color: #3a405a; -fx-border-width: 0px0px1.5px0px;" AnchorPane.leftAnchor="135.0" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="200.0">
          <font>
             <Font name="Microsoft Sans Serif" size="12.0" />
          </font>
@@ -50,15 +50,13 @@
             <Cursor fx:constant="HAND" />
          </cursor>
       </PasswordField>
-      <CheckBox fx:id="rememberDevice" layoutX="150.0" layoutY="243.0" mnemonicParsing="false" prefHeight="14.0" prefWidth="220.0" stylesheets="@../../../Stylesheets/Login.css" text="Keep me logged in" textFill="#50576d80" AnchorPane.leftAnchor="135.0" AnchorPane.rightAnchor="135.0" AnchorPane.topAnchor="235.0">
-         <font>
-            <Font name="Microsoft Sans Serif" size="8.0" />
-         </font>
-      </CheckBox>
       <ImageView fitHeight="66.0" fitWidth="75.0" layoutX="227.0" layoutY="15.0" pickOnBounds="true" preserveRatio="true" style="-fx-cursor: hand;" AnchorPane.leftAnchor="227.0" AnchorPane.rightAnchor="227.0" AnchorPane.topAnchor="35.0">
          <image>
             <Image url="@../../../Assets/AppIcon.png" />
          </image>
+         <effect>
+            <ColorAdjust brightness="-0.25" />
+         </effect>
       </ImageView>
       <Label fx:id="loginMessage" alignment="CENTER" layoutX="10.0" layoutY="153.0" textAlignment="CENTER" textFill="RED" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="140.0">
          <font>
@@ -71,5 +69,6 @@
          </font>
       </Button>
       <Button fx:id="minimizeButton" layoutX="235.0" layoutY="10.0" mnemonicParsing="false" onAction="#minimizeAction" prefHeight="30.0" prefWidth="30.0" stylesheets="@../../../Stylesheets/Login.css" text="➖" textAlignment="CENTER" AnchorPane.rightAnchor="30.0" AnchorPane.topAnchor="0.0" />
+      <Button fx:id="adminButton" layoutX="215.0" layoutY="361.0" mnemonicParsing="false" onAction="#switchToAdminLogin" style="-fx-background-color: transparent;" text="Admin sign-in" textFill="#00000080" />
    </children>
 </AnchorPane>
