+++ src/main/resources/com/example/orderly_inventory_management/StudentSignOutScene.fxml
@@ -1,0 +1,101 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.effect.ColorAdjust?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.shape.Line?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane prefHeight="768.0" prefWidth="1366.0" style="-fx-background-color: #101010;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.StudentSignOutController">
+   <children>
+      <AnchorPane layoutX="5.0" layoutY="45.0" prefHeight="719.0" prefWidth="1304.0" style="-fx-background-color: #1c1c1c; -fx-background-radius: 8;" AnchorPane.bottomAnchor="5.0" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="310.0" AnchorPane.topAnchor="50.0">
+         <children>
+            <TableView fx:id="table_students" editable="true" layoutX="-115.0" layoutY="-26.0" prefHeight="713.0" prefWidth="1015.0" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="45.0">
+               <columns>
+                  <TableColumn fx:id="col_name" prefWidth="75.0" text="Name" />
+                  <TableColumn fx:id="col_item" prefWidth="75.0" text="Item" />
+                  <TableColumn fx:id="col_amount" prefWidth="75.0" text="Amount" />
+                  <TableColumn fx:id="col_returned" prefWidth="75.0" text="Returned" />
+                  <TableColumn fx:id="col_date" prefWidth="75.0" text="Date Checked Out" />
+               </columns>
+               <opaqueInsets>
+                  <Insets />
+               </opaqueInsets>
+               <columnResizePolicy>
+                  <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+               </columnResizePolicy>
+            </TableView>
+            <TextField fx:id="keywordTextField" layoutX="12.0" layoutY="16.0" prefHeight="25.0" promptText="Search" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.leftAnchor="10.0" AnchorPane.topAnchor="10.0">
+               <font>
+                  <Font size="10.0" />
+               </font>
+            </TextField>
+            <Line endX="-100.0" endY="24.5" layoutX="249.0" layoutY="-13.0" startX="-100.0" startY="45.0" stroke="#989898" strokeWidth="2.0" AnchorPane.leftAnchor="147.0" AnchorPane.topAnchor="10.5" />
+            <Label layoutX="158.0" layoutY="12.0" text="Sort by" textFill="WHITE" AnchorPane.leftAnchor="160.0" AnchorPane.topAnchor="14.0" />
+            <ComboBox layoutX="246.0" layoutY="9.0" maxHeight="-Infinity" minHeight="-Infinity" prefHeight="25.0" prefWidth="100.0" promptText="..." stylesheets="@../../../Stylesheets/Global.css" AnchorPane.leftAnchor="205.0" AnchorPane.topAnchor="10.0">
+               <effect>
+                  <ColorAdjust />
+               </effect>
+            </ComboBox>
+         </children>
+      </AnchorPane>
+      <AnchorPane fx:id="topBar" layoutX="5.0" layoutY="7.0" prefHeight="40.0" prefWidth="1356.0" style="-fx-background-color: #1C1C1C; -fx-background-radius: 8;" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="5.0">
+         <children>
+            <Button fx:id="closeButton" layoutX="1297.0" layoutY="8.0" maxHeight="-Infinity" minHeight="-Infinity" mnemonicParsing="false" onAction="#quitOnAction" prefHeight="34.0" prefWidth="34.0" stylesheets="@../../../Stylesheets/Global.css" text="❌" textFill="WHITE" AnchorPane.bottomAnchor="3.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="3.0">
+               <font>
+                  <Font name="System Bold" size="12.0" />
+               </font>
+            </Button>
+            <Button fx:id="minimizeButton" layoutX="1262.0" layoutY="8.0" mnemonicParsing="false" onAction="#minimizeAction" prefHeight="34.0" prefWidth="34.0" stylesheets="@../../../Stylesheets/Global.css" text="−" textFill="WHITE" AnchorPane.bottomAnchor="3.0" AnchorPane.rightAnchor="95.0" AnchorPane.topAnchor="3.0">
+               <font>
+                  <Font name="System Bold" size="12.0" />
+               </font>
+            </Button>
+            <ImageView fitHeight="55.0" fitWidth="30.0" layoutX="14.0" layoutY="5.0" pickOnBounds="true" preserveRatio="true" AnchorPane.leftAnchor="5.0">
+               <image>
+                  <Image url="@../../../Assets/AppIcon.png" />
+               </image>
+               <effect>
+                  <ColorAdjust brightness="1.0" />
+               </effect>
+            </ImageView>
+            <Label alignment="BASELINE_LEFT" layoutX="703.0" layoutY="7.0" text="ORDERLY" textFill="WHITE" AnchorPane.leftAnchor="40.0">
+               <font>
+                  <Font name="System Bold" size="18.0" />
+               </font>
+            </Label>
+            <Button fx:id="fullScreenButton" layoutX="1222.0" maxHeight="-Infinity" minHeight="-Infinity" mnemonicParsing="false" onAction="#fullScreenAction" prefHeight="34.0" prefWidth="34.0" stylesheets="@../../../Stylesheets/Global.css" text="❐" textFill="WHITE" AnchorPane.bottomAnchor="3.0" AnchorPane.rightAnchor="55.0" AnchorPane.topAnchor="3.0">
+               <font>
+                  <Font name="System Bold" size="12.0" />
+               </font>
+            </Button>
+         </children>
+      </AnchorPane>
+      <AnchorPane layoutX="1062.0" layoutY="50.0" prefHeight="713.0" prefWidth="300.0" style="-fx-background-color: #1c1c1c; -fx-background-radius: 8;" AnchorPane.bottomAnchor="5.0" AnchorPane.leftAnchor="1061.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="50.0">
+         <children>
+            <Label alignment="CENTER" layoutX="87.0" layoutY="49.0" text="Unretruned Items" textFill="WHITE" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="18.0">
+               <font>
+                  <Font name="System Bold" size="18.0" />
+               </font>
+            </Label>
+            <TableView editable="true" layoutX="50.0" layoutY="110.0" prefHeight="200.0" prefWidth="200.0" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="65.0">
+              <columns>
+                <TableColumn prefWidth="240.0" text="Item" />
+                <TableColumn prefWidth="60.0" />
+              </columns>
+               <columnResizePolicy>
+                  <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+               </columnResizePolicy>
+            </TableView>
+         </children>
+      </AnchorPane>
+   </children>
+</AnchorPane>
