+++ src/main/resources/com/example/orderly_inventory_management/TransactionScene.fxml
@@ -8,6 +8,7 @@
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.effect.ColorAdjust?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -15,7 +16,7 @@
 <?import javafx.scene.shape.Line?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane fx:id="parent" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="768.0" prefWidth="1366.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.TransactionSceneController">
+<AnchorPane fx:id="parent" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="768.0" prefWidth="1366.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.TransactionController">
     <children>
         <BorderPane style="-fx-background-color: #101010;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
             <right>
@@ -31,7 +32,7 @@
                                         <Line endX="280.0" startX="16.0" stroke="#989898" strokeWidth="2.0" AnchorPane.topAnchor="60.0" />
                               <Label alignment="CENTER" layoutX="95.0" layoutY="15.0" stylesheets="@../../../Stylesheets/Global.css" text="MODIFICATIONS" textFill="WHITE" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="18.0">
                                  <font>
-                                    <Font name="System Bold Italic" size="18.0" />
+                                    <Font name="System Bold" size="18.0" />
                                  </font>
                               </Label>
                                     </children>
@@ -124,7 +125,6 @@
                             </cursor>
                         </Button>
                         <Button fx:id="transactionButton" alignment="BASELINE_LEFT" layoutX="15.0" layoutY="233.0" mnemonicParsing="false" onAction="#switchToTransaction" prefHeight="50.0" stylesheets="@../../../Stylesheets/Transaction.css" text="Transactions" textFill="#6c757d" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="160.0">
-                        <Button fx:id="transactionButton" alignment="BASELINE_LEFT" layoutX="15.0" layoutY="233.0" mnemonicParsing="false" onAction="#switchToTransaction" prefHeight="50.0" style="-fx-background-color: #3d3d3d; -fx-background-radius: 5; -fx-text-fill: white;" text="Transactions" textFill="#6c757d" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="160.0">
                             <font>
                                 <Font name="System Bold" size="18.0" />
                             </font>
@@ -142,14 +142,6 @@
                             </cursor>
                         </Button>
                         <Button fx:id="dashboardButton" alignment="BASELINE_LEFT" layoutX="15.0" layoutY="133.0" mnemonicParsing="false" onAction="#switchToDashboard" prefHeight="50.0" stylesheets="@../../../Stylesheets/Global.css" text="Dashboard" textFill="#6c757d" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="60.0">
-                            <font>
-                                <Font name="System Bold" size="18.0" />
-                            </font>
-                            <cursor>
-                                <Cursor fx:constant="HAND" />
-                            </cursor>
-                        </Button>
-                        <Button fx:id="notiButton" alignment="BASELINE_LEFT" layoutX="15.0" layoutY="283.0" mnemonicParsing="false" prefHeight="50.0" stylesheets="@../../../Stylesheets/Global.css" text="Notifications" textFill="#6c757d" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="210.0">
                             <font>
                                 <Font name="System Bold" size="18.0" />
                             </font>
@@ -204,6 +196,9 @@
                             <image>
                                 <Image url="@../../../Assets/AppIcon.png" />
                             </image>
+                     <effect>
+                        <ColorAdjust brightness="1.0" />
+                     </effect>
                         </ImageView>
                         <Label alignment="BASELINE_LEFT" layoutX="703.0" layoutY="7.0" text="ORDERLY" textFill="WHITE" AnchorPane.leftAnchor="40.0">
                             <font>
@@ -226,12 +221,12 @@
                     <children>
                         <TableView fx:id="table_items" editable="true" layoutX="208.0" layoutY="157.0" prefHeight="200.0" prefWidth="200.0" stylesheets="@../../../Stylesheets/Global.css" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="45.0">
                             <columns>
-                                <TableColumn fx:id="col_component" prefWidth="75.0" text="Component" />
-                                <TableColumn fx:id="col_value" prefWidth="75.0" text="Value" />
-                                <TableColumn fx:id="col_amount" prefWidth="75.0" text="Amount" />
-                                <TableColumn fx:id="col_dlb" prefWidth="75.0" text="Date Last Bought" />
-                                <TableColumn fx:id="col_link" prefWidth="75.0" text="Link" />
-                                <TableColumn fx:id="col_select" prefWidth="75.0" text="Select" />
+                                <TableColumn fx:id="col_component" prefWidth="156.0" text="Component" />
+                                <TableColumn fx:id="col_value" prefWidth="156.0" text="Value" />
+                                <TableColumn fx:id="col_amount" prefWidth="156.0" text="Amount" />
+                                <TableColumn fx:id="col_dlb" prefWidth="156.0" text="Date Last Bought" />
+                                <TableColumn fx:id="col_link" prefWidth="156.0" text="Link" />
+                                <TableColumn fx:id="col_select" prefWidth="64.0" text="Select" />
                             </columns>
                             <columnResizePolicy>
                                 <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
